[#122] refactor: contract command 제거
