+++ src/main/java/org/silsagusi/joonggaemoa/domain/contract/service/dto/ContractSummaryResponse.java
@@ -1,4 +1,4 @@
-package org.silsagusi.joonggaemoa.domain.contract.controller.dto;
+package org.silsagusi.joonggaemoa.domain.contract.service.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class ContractSummaryResponse {
 
-	private Long count;
-	private Double rate;
+    private Long count;
+    private Double rate;
 }
